+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # CHANGELOG
 
 ## version 6.0.0 (unreleased)
-- **endianness** as a ES6 module: endianness.js
-- **"main"** dist file is a UMD: endianness.umd.js
+- **endianness** as a ES6 module
+- Option to swap only a slice of the buffer